+++ src/main/java/de/elomagic/spps/bc/SimpleCrypt.java
@@ -40,7 +40,6 @@ import java.nio.file.FileAlreadyExistsException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.security.GeneralSecurityException;
 import java.security.Key;
 import java.security.SecureRandom;
 import java.util.Arrays;
@@ -83,22 +82,22 @@ public class SimpleCrypt {
      * Read private key from default location.
      *
      * @return Returns the private key.
-     * @throws GeneralSecurityException Thrown when unable to create private key
+     * @throws SimpleCryptException Thrown when unable to create private key
      */
     @NotNull
-    private static Key readPrivateKey() throws GeneralSecurityException {
+    private static Key readPrivateKey() throws SimpleCryptException {
         return readPrivateKey(SETTINGS_FILE.get());
     }
 
     /**
-     * Read a private key.
+     * Read a private key file.
      *
      * @param file File of the private key. When relocation in file is set then key will be read from there.
      * @return Returns the private key.
-     * @throws GeneralSecurityException Thrown when unable to create private key
+     * @throws SimpleCryptException Thrown when unable to create private key
      */
     @NotNull
-    private static Key readPrivateKey(@NotNull Path file) throws GeneralSecurityException {
+    private static Key readPrivateKey(@NotNull Path file) throws SimpleCryptException {
         try {
             if (Files.notExists(file)) {
                 throw new FileNotFoundException("Unable to find settings file. At first you have to create a private key.");
@@ -113,7 +112,7 @@ public class SimpleCrypt {
                 } else {
                     String key = p.getProperty(KEY_KEY, "");
                     if ("".equals(key)) {
-                        throw new GeneralSecurityException("No private key set.");
+                        throw new SimpleCryptException("No private key set.");
                     }
                     byte[] result = Base64.decode(key);
                     return new SecretKeySpec(result, ALGORITHM);
@@ -121,19 +120,19 @@ public class SimpleCrypt {
             }
         } catch (Exception ex) {
             LOGGER.error(ex.getMessage(), ex);
-            throw new GeneralSecurityException("Unable to read private key.", ex);
+            throw new SimpleCryptException("Unable to read private key.", ex);
         }
     }
 
     /**
      * Creates a private key file.
      *
-     * @param file Private key file
+     * @param file File where the private key will be stored. If null then default file, which be stored in the user folder, will be used.
      * @param relocationFile Alternative file where to write file with private key
      * @param force When true and private key file already exists then it will be overwritten otherwise an exception will be thrown
-     * @throws GeneralSecurityException Thrown when unable to create private key
+     * @throws SimpleCryptException Thrown when unable to create private key
      */
-    static void createPrivateKey(@Nullable Path file, @Nullable Path relocationFile, boolean force) throws GeneralSecurityException {
+    static void createPrivateKey(@Nullable Path file, @Nullable Path relocationFile, boolean force) throws SimpleCryptException {
         try {
             file = file == null ? SETTINGS_FILE.get() : file;
 
@@ -165,6 +164,8 @@ public class SimpleCrypt {
                 createPrivateKey(relocationFile, null, force);
             }
 
+            LOGGER.info("Creating settings file");
+
             try (Writer writer = Files.newBufferedWriter(file, StandardCharsets.UTF_8)) {
                 p.store(writer, "SPPS Settings");
             }
@@ -173,7 +174,7 @@ public class SimpleCrypt {
             throw new IllegalStateException("Unable to create private key", ex);
         } catch (Exception ex) {
             LOGGER.error(ex.getMessage(), ex);
-            throw new GeneralSecurityException("Unable to create or read private key.", ex);
+            throw new SimpleCryptException("Unable to create or read private key.", ex);
         }
     }
 
@@ -185,22 +186,54 @@ public class SimpleCrypt {
      * @return Returns cipher
      */
     @NotNull
-    private static Cipher createCypher(int opmode, @NotNull IvParameterSpec iv) throws GeneralSecurityException {
-        Cipher cipher = Cipher.getInstance(TRANSFORMATION, new BouncyCastleProvider());
-        cipher.init(opmode, readPrivateKey(), iv);
+    private static Cipher createCypher(int opmode, @NotNull IvParameterSpec iv) throws SimpleCryptException {
+        try {
+            Cipher cipher = Cipher.getInstance(TRANSFORMATION, new BouncyCastleProvider());
+            cipher.init(opmode, readPrivateKey(), iv);
 
-        return cipher;
+            return cipher;
+        } catch (Exception ex) {
+            throw new SimpleCryptException(ex.getMessage(), ex);
+        }
+    }
+
+    /**
+     * Checks if settings file exists.
+     *
+     * Settings file can be "${user.home}/.spps/settings" or an alternative file which set with {@link SimpleCrypt#setSettingsFile(Path)}
+     *
+     * @return Returns true if exists.
+     */
+    static boolean isInitialize() {
+        return Files.exists(SETTINGS_FILE.get());
+    }
+
+    /**
+     * Initialize SimpleCrypt by checking of setting file.
+     * <p>
+     * Settings file can be "${user.home}/.spps/settings" or an alternative file which set with {@link SimpleCrypt#setSettingsFile(Path)}
+     *
+     * @throws SimpleCryptException Thrown when unable to create private key file
+     * @return Returns true when settings file was created and false when settings file already exist.
+     */
+    public static boolean init() throws SimpleCryptException {
+        if (!isInitialize()) {
+            createPrivateKey(null, null, false);
+            return true;
+        }
+
+        return false;
     }
 
     /**
      * Encrypt, encoded as Base64 and encapsulate with curly bracket of a string.
      *
      * @param decrypted a non encrypted byte array
-     * @return Returns a encrypted, Base64 encoded string, surrounded with curly brackets.
-     * @throws GeneralSecurityException Thrown when an error occurred during encrypting.
+     * @return Returns an encrypted, Base64 encoded string, surrounded with curly brackets.
+     * @throws SimpleCryptException Thrown when an error occurred during encrypting.
      */
     @Nullable
-    public static String encrypt(@Nullable byte[] decrypted) throws GeneralSecurityException {
+    public static String encrypt(byte[] decrypted) throws SimpleCryptException {
         if (decrypted == null) {
             return null;
         }
@@ -217,7 +250,7 @@ public class SimpleCrypt {
             return "{" + Base64.toBase64String(data) + "}";
         } catch(Exception ex) {
             LOGGER.error(ex.getMessage(), ex);
-            throw new GeneralSecurityException(ex.getMessage(), ex);
+            throw new SimpleCryptException(ex.getMessage(), ex);
         }
     }
 
@@ -225,11 +258,11 @@ public class SimpleCrypt {
      * Encrypt, encode as Base64 and encapsulate with curly bracket of a string.
      *
      * @param decrypted a non encrypted char array
-     * @return Returns a encrypted, Base64 encoded string, surrounded with curly brackets.
-     * @throws GeneralSecurityException Thrown when an error occurred during encrypting.
+     * @return Returns an encrypted, Base64 encoded string, surrounded with curly brackets.
+     * @throws SimpleCryptException Thrown when an error occurred during encrypting.
      */
     @Nullable
-    public static String encrypt(@Nullable char[] decrypted) throws GeneralSecurityException {
+    public static String encrypt(char[] decrypted) throws SimpleCryptException {
         return decrypted == null ? null : encrypt(CharUtils.toByteArray(decrypted));
     }
 
@@ -237,11 +270,11 @@ public class SimpleCrypt {
      * Encrypt, encode as Base64 and encapsulate with curly bracket of a string.
      *
      * @param decrypted a non encrypted string
-     * @return Returns a encrypted, Base64 encoded string, surrounded with curly brackets.
-     * @throws GeneralSecurityException Thrown when an error occurred during encrypting.
+     * @return Returns an encrypted, Base64 encoded string, surrounded with curly brackets.
+     * @throws SimpleCryptException Thrown when an error occurred during encrypting.
      */
     @Nullable
-    public static String encrypt(@Nullable String decrypted) throws GeneralSecurityException {
+    public static String encrypt(@Nullable String decrypted) throws SimpleCryptException {
         return decrypted == null ? null : encrypt(decrypted.getBytes(StandardCharsets.UTF_8));
     }
 
@@ -250,16 +283,16 @@ public class SimpleCrypt {
      *
      * @param encryptedBase64 Base64 encoded data string, encapsulated with curly brackets.
      * @return The encrypted data as string.
-     * @throws GeneralSecurityException Thrown when unable to decrypt data .
+     * @throws SimpleCryptException Thrown when unable to decrypt data .
      */
     @Nullable
-    public static byte[] decrypt(@Nullable String encryptedBase64) throws GeneralSecurityException {
+    public static byte[] decrypt(@Nullable String encryptedBase64) throws SimpleCryptException {
         if (encryptedBase64 == null) {
             return null;
         }
 
         if(!isEncryptedValue(encryptedBase64)) {
-            throw new GeneralSecurityException("This value is not with curly brackets encapsulated as an encrypted value. Unable to decrypt.");
+            throw new SimpleCryptException("This value is not with curly brackets encapsulated as an encrypted value. Unable to decrypt.");
         }
 
         try {
@@ -272,7 +305,7 @@ public class SimpleCrypt {
             return cipher.doFinal(encryptedBytes, 16, encryptedBytes.length-16);
         } catch(Exception ex) {
             LOGGER.error(ex.getMessage(), ex);
-            throw new GeneralSecurityException("Unable to decrypt data.", ex);
+            throw new SimpleCryptException("Unable to decrypt data.", ex);
         }
     }
 
@@ -281,10 +314,10 @@ public class SimpleCrypt {
      *
      * @param encryptedBase64 Base64 encoded data string, encapsulated with curly brackets.
      * @return The encrypted data as char array.
-     * @throws GeneralSecurityException Thrown when unable to decrypt data .
+     * @throws SimpleCryptException Thrown when unable to decrypt data .
      */
     @Nullable
-    public static char[] decryptToChars(@Nullable String encryptedBase64) throws GeneralSecurityException {
+    public static char[] decryptToChars(@Nullable String encryptedBase64) throws SimpleCryptException {
         return encryptedBase64 == null ? null : ByteUtils.toCharArray(decrypt(encryptedBase64));
     }
 
@@ -293,10 +326,10 @@ public class SimpleCrypt {
      *
      * @param encryptedBase64 Base64 encoded data string, encapsulated with curly brackets.
      * @return The encrypted data as string.
-     * @throws GeneralSecurityException Thrown when unable to decrypt data .
+     * @throws SimpleCryptException Thrown when unable to decrypt data .
      */
     @Nullable
-    public static String decryptToString(@Nullable String encryptedBase64) throws GeneralSecurityException {
+    public static String decryptToString(@Nullable String encryptedBase64) throws SimpleCryptException {
         // For JUnit test we have to use System.out because console() will return null
         return encryptedBase64 == null ? null : new String(decrypt(encryptedBase64), StandardCharsets.UTF_8);
     }