+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [1.2.0] - Upcoming
+
+* FEATURE Method "init()" added to create settings file if it doesn't exist. Recommend to call during application start.
+
+### Upgrade Notes:
+* Exception class GeneralSecurityException replaced by runtime exception SimpleCryptException
+
 ## [1.1.2] - 2021-12-17
 * BUGFIX Log4shell issue fixed
 
@@ -7,7 +14,6 @@
 
 * BUGFIX Possible NPE occur on empty arguments 
 * FEATURE Set alternative default settings file
-
 
 ## [1.0.1] - 2021-03-08